requirements.txt
# Image AI Analysis - Requirements

# Core dependencies
streamlit>=1.47.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.53.0
pillow>=10.0.0

# Image processing
numpy>=1.23.0
opencv-python>=4.8.0

# Data analysis (for analysis.py)
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
yfinance>=0.2.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
prophet>=1.1.0

# Web interface
gradio>=4.0.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
packaging>=23.0
